refactor(assignment): stay within source budgets
